Rename ImageAutoResizer to MediaAutoResizer
This is a proactive change of the class name so that we could possibly
add video resizing support to it later, without breaking backwards
compatibility by suddenly changing its name later.

And we don't even need any new constructor params to tell it what type
of file it is... If we ever add video processing support, we could make
it auto-detect format based on filename extension (jpg, jpeg, gif, png,
bmp = photo, everything else = treat as video), and perhaps add a
$options['filetype'] to override the autodetection.

Anyway, these are future plans and may not happen. But this name change
is good anyway since all of the algorithms for aspect ratio/dimension
determining are the exact same for both videos and photos, and it'd be
nice to do it all with a single class.