Created pre-commit hook to prevent bad commits
Here we go! This is why I just rewrote the various codestyle checkers to
support a "validate-only" mode...

This is now a fully automated git commit hook which runs all codestyle
checks on the git worktree and refuses to commit if we have any problems
in our worktree. We all forget to run codestyle manually sometimes... ;)

So we must all install this hook into our repos now, to prevent anymore
sloppy commits. It's easy. Just run this command in your local repository:

```
ln -fs "../../.pre-commit.hook" .git/hooks/pre-commit
```

From then on, it will totally refuse to commit if there are problems.

@valga @mgp25 ;)