Utils: Rewrite the URL syntax validator
It turns out that the "filter_var" URL validator is extremely strict
and doesn't allow Unicode (it expects %FF-encoding of all non-ASCII
characters). It therefore won't work with non-English URLs.

So I rewrote the whole thing via a regex which verifies that we only
catch http/https URLs with what seems to be a valid web domain name,
and that it has no whitespace anywhere in the URL.