MediaAutoResizer: Add support for story media
There are quite a lot of changes, because stories are very complex:

Instagram stories uploaded from real devices are always within a certain
range of "normal" aspect ratios, but those ratios don't look good on
9:16 screens. The best ratio is 9:16 (0.5625).

This new implementation adds a "targetFeed" option to the MediaAutoResizer.
If set to any of the story feed constants, it uses "story media" logic
instead of "general/timeline media" logic.

There are two sets of this new story logic: "useBestStoryRatio" (on by
default), which only allows a very small range around ~0.5625 ratio
before deciding to resize the image (this is on by default so that it
auto-targets 9:16 by default). And with that option off, it allows the
full range of normal device story ratios (some of which will look very
bad and strangely zoomed on 16:9 screens).

Regardless of which range is used (regardless of the "useBestStoryRatio"
value), we will ALWAYS target 0.5625 ratio when the story media is outside
of the allowed range. That's because it's the only ratio that looks good
on the majority of device screens due to how stories are scaled/zoomed.

So... Just provide the new value, "targetFeed" => Constants::FEED_STORY
and it'll use the "target 9:16" logic by default. You can further tweak
that via "useBestStoryRatio" and "min/maxAspectRatio" parameters as usual.

Enjoy!

@valga There are lots of new constants now, and I've removed the
MAX_HEIGHT since it wasn't needed for anything and just complicated
calculations when we involve stories. As long as width is legal and
aspect is legal, we know that the height is 100% legal too.