StorageHandler: Easy system for resetting device
In the past, we needed to constantly maintain the list of old settings
to reset in setUser() when the user's Android device changes.

That was error-prone, since many contributors forgot to add new settings
to the reset-list, which would mean that there was a mix of old and new
settings in the storage. For example, old re-used hardware IDs together
with a new device, which could look very suspicious to Instagram.

This new system is much easier to maintain. There's now a simple list of
"settings-keys to keep when erasing current device" in StorageHandler,
and it's very visibly placed immediately after the list of all available
settings-keys. In that list, we simply have to add any keys that we
DON'T WANT TO LOSE when the user's Android device changes!

For now, we don't need to keep anything since ALL except ONE of the
settings are device-specific, so I added our one-and-only generic
settings key as an example!