Request: Fix login condition + Registration: Update functions
Request:

I have fixed the following condition:

 if (!$this->_parent->isLoggedIn && $this->_parent->username !==
Instagram::ANONYMOUS_USER)

It is a completely failure, and i’ll tell you why. After doing this, i
realized that an anonymous user could also make authenticated requests
and we have done something to deal with this, that is setNeedsAuth().

So basically i have fixed the condition to what it was before the
previous commits and the update the Registration functions with
setNeedsAuth().

Since a registered user can change its own username and email, these
function can also be called by authenticated users, thats why we check
if its the fictitious username by doing:

 $this->ig->username === \InstagramAPI\Instagram::ANONYMOUS_USER

In case it is the fictitious username, we simple setNeedsAuth(false);

If it is an authenticated user we need to add the post param '_uid'

Still improving my logic and coding thinking.

Sorry for the inconveniences.