Response/GenericResponse: Remove unwanted fields
These were recently added by mgp25 after a user wanted to be able to get
the feedback error details. But this is totally the wrong way to get
those details.

Error fields don't belong in response classes.

Here's how to get error fields (things which are not mapped in responses):
https://github.com/mgp25/Instagram-API/issues/1199#issuecomment-296448210

For example:
```
try {
    // post comment
} catch (\InstagramAPI\Exception\FeedbackRequiredException $e) {
    if ($e->hasResponse()) {
        var_dump($e->getResponse()->getFullResponse()->feedback_title);
    }
}
```