AutoPropertyHandler: Massive performance boost
This implements a lookup cache for the AutoPropertyHandler. At the cost
of a little bit of RAM (the cache can be disabled on low-RAM systems),
it improves the speed of function calls so that they only take ~16% of
the time they used to take before we implemented this caching!

Tip: If you have limited memory, you can easily disable this cache:
\InstagramAPI\AutoPropertyHandler::$useCache = false;