Request and Request/Business: New function to get account statistics
This commit adds a new function `getStatistics()` for business
profiles. With this function you can get account statistics and other
information.

With this commit i have edited `\Request` since the response of the
function was 2 different json. So a new setter for Request was added:
`setIsMultiResponse()` which is off by default. When
`setIsMultiResponse` is on, it will simply merge multiple responses in
to single one. It could be done in several ways (array_merge and so),
but @valga came with a good idea just replacing the ending of the
response and the beginning of the other response.

The rest of the files are just Responses and Models to manage
`getStatistics()` response.