Rename ImageResizer to PhotoResizer
Now that all Media-related file processing classes are under
InstagramAPI\Media, we ended up with this situation:

Dimensions.php
ImageResizer.php
MediaDetails.php
PhotoDetails.php
Rectangle.php
ResizerInterface.php
VideoDetails.php
VideoResizer.php

The eagle-eyed will notice that "ImageResizer" and "PhotoDetails" don't
match.

So we have to name them either "Image" or "Photo", but not both.

But because Instagram uses the word "Photo" (and "Video"), our code is
already full of hundreds of sentences with the word "photo" in it.

Therefore, that decision wins. Sure, not every image is a "photo" (which
means image of a real object, taken with a camera). But it doesn't
matter much. What matters more is that we follow Instagram's
terminology.

And since Instagram is only meant for photos (taken with cameras), it's
best to stick with "Photo" as our word for images.

So in conclusion... we now have a "PhotoResizer" class!

Voila, all Media-class names are now matching up nicely:

Dimensions.php
MediaDetails.php
PhotoDetails.php
PhotoResizer.php
Rectangle.php
ResizerInterface.php
VideoDetails.php
VideoResizer.php