Media: Support custom video thumbnail timestamps
Closes #1831

Massive thanks to @tje3d who started the process of making this change,
but it unfortunately became a bit messy by the end. But the fact that he
started the work is the only reason that this feature exists now! Also
thanks to @valga for discussion and great ideas!

This commit does the following changes:

- Refactored `InstagramVideo` class to support adding both input-file
  flags and output-file flags (instead of only output-file flags), and
  made those functions return arrays for easier handling. This change
  was necessary for fast thumbnail seeking. Otherwise (if `-ss` is
  specified after the input file), ffmpeg converts the whole file up
  until the desired timestamp which is a big waste of time. ;)

- The `Internal` class now gives the optional `thumbnail_timestamp`
  `$externalMetadata` value into the constructor of `InstagramThumbnail`
  if a custom timestamp has been provided by the user.

- `InstagramThumbnail` now supports a "thumbnailTimestamp" option (note
  the different spelling for our internal name), which is only applied
  if the target feed for the video is TIMELINE (exactly like the real
  app). The thumbnail class converts and validates the given timestamp
  and ensures it's within a legal range, and then provides 3 ways to use
  the timestamp after that: `getFile()` to generate a jpeg thumbnail
  file as usual (but with the custom timestamps now), or
  `getTimestamp()` to get the validated timestamp as a float (useful by
  valga when implementing `cover_frame_seek` server-side cover
  generation later), and `getTimestampString()` which gets it as a
  ffmpeg-style string instead.

For users:

Your custom thumbnail timestamps can only be provided for timeline
videos (`$ig->timeline->uploadVideo()`). Not for any other video types
and not for any albums. This is intentional to match the real app.

You can provide the timestamp value as a float or as a time string.

Examples (both use 1.5 seconds as the cover timestamp location):

```
$ig->timeline->uploadVideo('example.mp4', ['thumbnail_timestamp' => 1.5]);
$ig->timeline->uploadVideo('example.mp4', ['thumbnail_timestamp' => '00:00:01.500']);
```

If your timestamp position is longer than the video, the final frame of
the video is used. So you don't have to worry about that.