InstagramThumbnail: Use 1st valid frame on failure
This is based on discussion here:
https://github.com/mgp25/Instagram-API/commit/96b02ec73256c865cfc5930d8b629d8dc7762bda

Some videos don't start with a valid keyframe, so we can't seek to
`00:00:00.000`. And it's impossible to know where the first valid
keyframe will sit in a file (it varies), so we can't hardcode something
like "minimum: 00.120" (and such a hack would be ugly for all the
files (majority) that *do* support seeking to 00.000).

The solution is to look specifically for the "first frame is no
keyframe" error after the first attempt, and rerunning the thumbnail
extraction without the `-ss` seek-command, so that we extract the first
valid frame instead of the user's timestamp.

The result of this matches what the user would have expected, since it
gives them the earliest valid frame. And it also ensures we always
support extraction even at 00.000 for all files that do support it.