Account: Fix editProfile(), use endpoint username
Our `$ig->username` property is not 100% reliable (especially if we
someday add Facebook logins). So we shouldn't use it for purposes like
this again.

Now we match the app behavior, which calls the `accounts/current_user/`
endpoint to get the username to fill into the "username" property if no
other new name has been provided.