InstagramAPI: Mark in-dev version as v5
Due to backwards-compatibility breaking changes in some recent commits,
we must mark the next version as v5.

v5 being next of course means that the dev team is welcome to break
backwards compatibility as much as you want in `dev-master`. So to all
devs: Feel free to do any big library refactorings you've been dreaming of.